Split map module into multiple files.
